requirements.txt
Django==3.0.6
djangorestframework==3.11.0
django-model-utils==4.0.0
django-cors-headers==3.3.0
python-decouple==3.3
requests==2.23.0
celery==4.4.5
